+++ packages/build-tools/src/utils/__tests__/packageManager.test.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import { vol } from 'memfs';
 import fs from 'fs-extra';
 
-import { resolvePackageManager } from '../packageManager';
+import { resolvePackageManager, findPackagerRootDir } from '../packageManager';
 
 jest.mock('fs');
 
@@ -61,5 +61,94 @@ describe(resolvePackageManager, () => {
     await fs.writeFile(path.join(nestedDir, 'package.json'), 'content');
 
     expect(resolvePackageManager(nestedDir)).toBe('yarn');
+  });
+});
+
+describe(findPackagerRootDir, () => {
+  beforeEach(() => {
+    vol.reset();
+  });
+
+  it('returns the workspace root if the current dir is a workspace', async () => {
+    vol.fromJSON(
+      {
+        './package.json': JSON.stringify({
+          workspaces: ['some-package', 'react-native-project'],
+        }),
+        './some-package/package.json': JSON.stringify({
+          name: 'some-package',
+        }),
+        './react-native-project/package.json': JSON.stringify({
+          name: 'react-native-project',
+        }),
+      },
+      '/repo'
+    );
+
+    const rootDir = findPackagerRootDir('/repo/react-native-project');
+    expect(rootDir).toBe('/repo');
+  });
+
+  it(
+    `returns the current dir if it's not a workspace` +
+      ` (package.json exists in root dir and contains workspaces field)`,
+    async () => {
+      vol.fromJSON(
+        {
+          './package.json': JSON.stringify({
+            workspaces: ['some-package'],
+          }),
+          './some-package/package.json': JSON.stringify({
+            name: 'some-package',
+          }),
+          './react-native-project/package.json': JSON.stringify({
+            name: 'react-native-project',
+          }),
+        },
+        '/repo'
+      );
+
+      const rootDir = findPackagerRootDir('/repo/react-native-project');
+      expect(rootDir).toBe('/repo/react-native-project');
+    }
+  );
+
+  it(
+    `returns the current dir if it's not a workspace` +
+      ` (package.json exists in root dir and does not contain workspaces field)`,
+    async () => {
+      vol.fromJSON(
+        {
+          './package.json': JSON.stringify({}),
+          './some-package/package.json': JSON.stringify({
+            name: 'some-package',
+          }),
+          './react-native-project/package.json': JSON.stringify({
+            name: 'react-native-project',
+          }),
+        },
+        '/repo'
+      );
+
+      const rootDir = findPackagerRootDir('/repo/react-native-project');
+      expect(rootDir).toBe('/repo/react-native-project');
+    }
+  );
+
+  it(`returns the current dir if it's not a workspace (package.json does not exist in root dir) `, async () => {
+    vol.fromJSON(
+      {
+        './some-package/package.json': JSON.stringify({
+          name: 'some-package',
+        }),
+        './react-native-project/package.json': JSON.stringify({
+          name: 'react-native-project',
+        }),
+      },
+      '/repo'
+    );
+
+    const rootDir = findPackagerRootDir('/repo/react-native-project');
+    expect(rootDir).toBe('/repo/react-native-project');
   });
 });
